refactor: data_structuresパッケージにSymbolManagerの2つを移動
